Fixes missing class declaration for TaxonRank:TaxonRankTerm
Without this declaration, it defaults to a datatype, making the `tc:rank` property declaration become punned between an ObjectProperty (with a domain constraint and all the annotation) and a DatatypeProperty (with range constraint on `TaxonRankTerm`). While not only obviously undesired (and undesirable), this in turn will lead to a DL reasoner error on the undeclared datatype `TaxonRankTerm` (which was never meant to be a datatype), resulting in failure to include the ontology in anything that requires OWL reasoning.

An alternative for fixing the issue would be to `owl:imports` the TaxonRank vocabulary. However, I have to say I actually _appreciate_ that the TaxonConcept ontology doesn't import `Base`, `Common`, etc, and thus is nicely self-contained for those who just care about the taxon concept vocabulary.